Add Google's webgl-utils.
This script is licensed under the BSD 3-Clause License, which according to
http://opensource.org/licenses/BSD-3-Clause
"You may omit clause 3 and still be OSD-conformant. Despite its colloquial name "BSD New", this is not the newest version of the BSD license; it was followed by the even newer BSD-2-Clause version, sometimes known as the "Simplified BSD License". On January 9th, 2008 the OSI Board approved BSD-2-Clause, which is used by FreeBSD and others. It omits the final "no-endorsement" clause and is thus roughly equivalent to the MIT License."

This file contains a cross-browser way of initializing WebGL.