Small miscellaneous changes to shader utilities.
Most of these changes stem from an attempt to add asynchronous loading of shaders from files rather than have them embedded in the HTML. While the changes did in fact work, there was a potential race condition that, try as I might, I cannot seem to head-off. While I could leave it and just hope it never occurs, or that affected viewers would just refresh, good programming practice and Murphy's law suggests that I leave it until I can think of a good solution. The upside is that I don't have to fire up an HTTP deamon prior to working on it locally.

Regardless, commit the changes that affect the other parts of the code.