Due to JavaScript's single-threaded event-driven model, don't worry about races.
I read a number of articles regarding concurrency in JavaScript though the clearest appears to be this one (written today, no less) :
https://uzairfarooq.github.io/why-no-concurrency-control-tool-in-javascript/

Put simply a block of JavaScript always runs to completion.

I put a delay in the load_shaders function between loads, the idea being to cause the callback to trigger twice, and could not get it to break, so for now as long as load_shaders is used then fears about race conditions can be allayed. Other use should be considered as it arises.

As a word of caution, clicking elements in the webpage linked to JavaScript functions will cause them to trigger events once control is restored, this may have unintended consequences for example : after the delay had finished but before the callback, code was then executed that relied on the WebGLRenderingContext which hadn't been initialized leading to a non-fatal error. I'm currently ignoring it, though adding some code to prevent execution until everything is initialized may not be a bad idea.

Upon reflection, JavaScript's model makes perfect sense for browser use, though understandably makes programmers used to other models nervous.