Add, but disable, save / download option.
I tried this a few days ago after looking at p5.js but couldn't get the content of the canvas. Daniela Baron posted on the Coursera forums the solution which is to pass the {preserveDrawingBuffer: true} as argument when creating the WebGLRenderingContext at startup. Daniela's technique differs from mine, though I only seem to get good results in Chrome, so for now I'm leaving it disabled.