Week 5 starts with a new assignment. Go go gadget third dimension.
I am currently offline and cannot check details, check techniques etc.
This commit contains a rotating 3D object that was cribbed from something I picked up from the learningwebgl.com blog and adapted. The result is hardly amazing but it works for now and shows the direction to persue.

It has to be said that the breakthough came when using FireFox's shader editor with requestAnimFrame to provide live updates.