Move 3D shapes into a common directory under directory "common".
Today, in the next commit, I'll be creating Chapter 7 which builds on these objects, so best not to duplicate code whenever wherever possible.

The third assignment is due at the end of the week, the fourth assignment builds on the third with lighting. Hopefully I can get most of this assignment done today, I can't really spend more than a two to three days dedicated to it as I have other assignments to hand in this week and a job interview to prep for.