Entertain non-HTML5 viewers.
Not a lot of work was done during week eight, at least nothing new has been added or committed. I have just short of a week to turn out the next assignment so expect a large amount of commits this week. I do however have an assignment due for another course at the end of this week, things are getting tight.