Multiple light sources.
Based on an anonymous post to the course forums.

Care must be taken that the number of lights in the shader match that of the lights in the JavaScript code. Some of the light rendering boilerplate can probably be moved into the Light object, but for now I'm just happy it works.