Basic Phong Shading.
It's not quite right, especially when the light is at the extremes of the X and Y axes, but for the most part it is aesthetically better and smoother than the previous Gouraud shading, in fact it may be worth backporting some of the improvements that have been discovered from implementing Phong Shading.

The goal at the moment is to implment Bump Mapping which I am led to believe is processed in the Fragment Shader as is Phong Shading, so this is an important first step.

Unfortunately I've had to reduce the number of lights down to one, which is a bit rubbish as it's all or nothing. I think I can increase the number of lights but it is not a pressing concern for this assignment. I still, obviously, have much to learn.

I find working in the shaders hard as there is little in the way of debugging support, experience will obviously make me more comfortable with it.